Scribbles - Remove model-based scribbles methods from generic segmentation app (#346)
* rename coldstart to model-free


* remove model-based scribble methods from generic segmentation app


* fix modelfree scribbles task, and codeformat issues


* pass through scaling and spacing arguments to scribbles infertask


* move scribbles infer to ./monailabel/scribbles/infer.py


* add model-free scribbles to all sample-apps


* rename model-free --> HistogramBasedGraphCut


* fix args to scribbles infer, and pre tx scaling


* fix codeformat issue


* fix intensity_range arguments
