+++ sample-apps/deepgrow/main.py
@@ -16,6 +16,7 @@ from lib import InferDeepgrow, MyStrategy, TrainDeepgrow
 from monai.apps import load_from_mmar
 
 from monailabel.interfaces.app import MONAILabelApp
+from monailabel.scribbles.infer import HistogramBasedGraphCut
 from monailabel.utils.activelearning.random import Random
 from monailabel.utils.infer.deepgrow_pipeline import InferDeepgrowPipeline
 
@@ -54,6 +55,7 @@ class MyApp(MONAILabelApp):
                 dimension=3,
                 model_size=(128, 192, 192),
             ),
+            "histogramBasedGraphCut": HistogramBasedGraphCut(),
         }
 
         infers["deepgrow_pipeline"] = InferDeepgrowPipeline(
