+++ sample-apps/segmentation/lib/__init__.py
@@ -11,11 +11,4 @@
 
 from .activelearning import MyStrategy
 from .infer import MyInfer
-from .scribbles import (
-    GenericInteractiveGraphCut,
-    GenericISegCRF,
-    GenericISegGraphCut,
-    GenericISegGraphcutColdstart,
-    GenericISegSimpleCRF,
-)
 from .train import MyTrain
