+++ sample-apps/segmentation_spleen_scribbles/lib/scribbles.py
@@ -13,13 +13,11 @@ from monai.transforms import AddChanneld, Compose, LoadImaged, ScaleIntensityRan
 
 from monailabel.interfaces.tasks.infer import InferTask, InferType
 from monailabel.scribbles.transforms import (
-    AddBackgroundScribblesFromROId,
     ApplyCRFOptimisationd,
     ApplyGraphCutOptimisationd,
     ApplyISegGraphCutPostProcd,
     ApplySimpleCRFOptimisationd,
     MakeISegUnaryd,
-    MakeLikelihoodFromScribblesHistogramd,
     SoftenProbSoftmax,
 )
 from monailabel.utils.others.post import BoundingBoxd, Restored
@@ -34,10 +32,14 @@ class SpleenPostProc(InferTask):
         self,
         dimension,
         description,
+        intensity_range=(-300, 200, 0.0, 1.0),
+        pix_dim=(2.5, 2.5, 5.0),
     ):
         super().__init__(
             path=None, network=None, labels=None, type=InferType.SCRIBBLES, dimension=dimension, description=description
         )
+        self.intensity_range = intensity_range
+        self.pix_dim = pix_dim
 
     def pre_transforms(self):
         return [
@@ -45,9 +47,15 @@ class SpleenPostProc(InferTask):
             AddChanneld(keys=["image", "label"]),
             # at the moment optimisers are bottleneck taking a long time,
             # therefore scaling non-isotropic with big spacing
-            Spacingd(keys=["image", "logits"], pixdim=[2.5, 2.5, 5.0]),
-            Spacingd(keys=["label"], pixdim=[2.5, 2.5, 5.0], mode="nearest"),
-            ScaleIntensityRanged(keys="image", a_min=-300, a_max=200, b_min=0.0, b_max=1.0, clip=True),
+            Spacingd(keys=["image", "logits", "label"], pixdim=self.pix_dim, mode=["bilinear", "bilinear", "nearest"]),
+            ScaleIntensityRanged(
+                keys="image",
+                a_min=self.intensity_range[0],
+                a_max=self.intensity_range[1],
+                b_min=self.intensity_range[2],
+                b_max=self.intensity_range[3],
+                clip=True,
+            ),
         ]
 
     def post_transforms(self):
@@ -79,8 +87,10 @@ class SpleenISegCRF(SpleenPostProc):
         self,
         dimension=3,
         description="A post processing step with ISeg + MONAI's CRF for Spleen segmentation",
+        intensity_range=(-300, 200, 0.0, 1.0),
+        pix_dim=(2.5, 2.5, 5.0),
     ):
-        super().__init__(dimension, description)
+        super().__init__(dimension, description, intensity_range, pix_dim)
 
     def pre_transforms(self):
         return [
@@ -88,9 +98,15 @@ class SpleenISegCRF(SpleenPostProc):
             AddChanneld(keys=["image", "label"]),
             # at the moment optimisers are bottleneck taking a long time,
             # therefore scaling non-isotropic with big spacing
-            Spacingd(keys=["image", "logits"], pixdim=[2.5, 2.5, 5.0]),
-            Spacingd(keys=["label"], pixdim=[2.5, 2.5, 5.0], mode="nearest"),
-            ScaleIntensityRanged(keys="image", a_min=-300, a_max=200, b_min=0.0, b_max=1.0, clip=True),
+            Spacingd(keys=["image", "logits", "label"], pixdim=self.pix_dim, mode=["bilinear", "bilinear", "nearest"]),
+            ScaleIntensityRanged(
+                keys="image",
+                a_min=self.intensity_range[0],
+                a_max=self.intensity_range[1],
+                b_min=self.intensity_range[2],
+                b_max=self.intensity_range[3],
+                clip=True,
+            ),
             SoftenProbSoftmax(logits="logits", prob="prob"),
         ]
 
@@ -108,68 +124,6 @@ class SpleenISegCRF(SpleenPostProc):
                 ),
                 # optimiser
                 ApplyCRFOptimisationd(unary="unary", pairwise="image", post_proc_label="pred", device="cpu"),
-            ]
-        )
-
-
-class SpleenISegGraphcutColdstart(SpleenPostProc):
-    """
-    Defines ISeg+Graphcut based cold start task for Spleen segmentation from the following paper:
-
-    Wang, Guotai, et al. "Interactive medical image segmentation using deep learning with image-specific fine tuning."
-    IEEE transactions on medical imaging 37.7 (2018): 1562-1573. (preprint: https://arxiv.org/pdf/1710.04043.pdf)
-
-    This task takes as input 1) original image volume and 2) scribbles from user
-    indicating foreground and background regions. A likelihood volume is generated using histogram method.
-    User-scribbles are incorporated using Equation 7 on page 4 of the paper.
-
-    SimpleCRF's Graphcut layer is used to optimise Equation 5 from the paper, where unaries come from Equation 7
-    and pairwise is the original input volume.
-    """
-
-    def __init__(
-        self,
-        dimension=3,
-        description="A post processing step with ISeg + Graphcut cold start for Spleen segmentation",
-    ):
-        super().__init__(dimension, description)
-
-    def pre_transforms(self):
-        return [
-            LoadImaged(keys=["image", "label"]),
-            AddChanneld(keys=["image", "label"]),
-            AddBackgroundScribblesFromROId(scribbles="label", scribbles_bg_label=2, scribbles_fg_label=3),
-            # at the moment optimisers are bottleneck taking a long time,
-            # therefore scaling non-isotropic with big spacing
-            Spacingd(keys=["image"], pixdim=[2.5, 2.5, 5.0]),
-            Spacingd(keys=["label"], pixdim=[2.5, 2.5, 5.0], mode="nearest"),
-            ScaleIntensityRanged(keys="image", a_min=-300, a_max=200, b_min=0.0, b_max=1.0, clip=True),
-            # SoftenProbSoftmax(logits="logits", prob="prob"),
-            MakeLikelihoodFromScribblesHistogramd(
-                image="image", scribbles="label", post_proc_label="prob", scribbles_bg_label=2, scribbles_fg_label=3
-            ),
-        ]
-
-    def inferer(self):
-        return Compose(
-            [
-                # unary term maker
-                MakeISegUnaryd(
-                    image="image",
-                    logits="prob",
-                    scribbles="label",
-                    unary="unary",
-                    scribbles_bg_label=2,
-                    scribbles_fg_label=3,
-                ),
-                # optimiser
-                ApplyGraphCutOptimisationd(
-                    unary="unary",
-                    pairwise="image",
-                    post_proc_label="pred",
-                    lamda=1.0,
-                    sigma=0.1,
-                ),
             ]
         )
 
@@ -193,8 +147,10 @@ class SpleenISegGraphCut(SpleenPostProc):
         self,
         dimension=3,
         description="A post processing step with ISeg + SimpleCRF's GraphCut for Spleen segmentation",
+        intensity_range=(-300, 200, 0.0, 1.0),
+        pix_dim=(2.5, 2.5, 5.0),
     ):
-        super().__init__(dimension, description)
+        super().__init__(dimension, description, intensity_range, pix_dim)
 
     def inferer(self):
         return Compose(
@@ -239,8 +195,10 @@ class SpleenInteractiveGraphCut(SpleenPostProc):
         self,
         dimension=3,
         description="A post processing step with SimpleCRF's Interactive ISeg GraphCut for Spleen segmentation",
+        intensity_range=(-300, 200, 0.0, 1.0),
+        pix_dim=(2.5, 2.5, 5.0),
     ):
-        super().__init__(dimension, description)
+        super().__init__(dimension, description, intensity_range, pix_dim)
 
     def inferer(self):
         return Compose(
@@ -278,8 +236,10 @@ class SpleenISegSimpleCRF(SpleenPostProc):
         self,
         dimension=3,
         description="A post processing step with ISeg + SimpleCRF's CRF for Spleen segmentation",
+        intensity_range=(-300, 200, 0.0, 1.0),
+        pix_dim=(2.5, 2.5, 5.0),
     ):
-        super().__init__(dimension, description)
+        super().__init__(dimension, description, intensity_range, pix_dim)
 
     def inferer(self):
         return Compose(
